backend/requirements.txt
python-dotenv
Django==6.0.2
asgiref==3.11.1
sqlparse==0.5.5
